requirements.txt
eventlet==0.32.0
python-socketio==5.4.1
redis==3.5.3
loguru==0.5.3
